pyproject.toml
[tool.poetry]
name = "blanche"
version = "1.2.1"
description = "The web browser for LLMs"
authors = ["Blanche Team <hello@blanche.cc>"]
readme = "README.md"
packages = [
  { include = "blanche" }
]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
playwright = "^1.48.0"
clean-html-for-llm = "^1.3.2"
webdriver-manager = "^4.0.2"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
bs4 = "^0.0.2"
loguru = "^0.7.2"
litellm = "^1.53.3"
chevron = "^0.14.0"
llamux = "^0.1.9"
pillow = "^11.1.0"

[tool.poetry.group.embeddings]
optional = true

[tool.poetry.group.embeddings.dependencies]
sentence-transformers = "^3.3.1"
numpy = "^2.2.1"

[tool.poetry.group.dev.dependencies]
types-requests = "^2.32.0.20241016"
mypy = "^1.13.0"
jupyter = "^1.1.1"
djlint = "^1.35.4"
pytest = "^8.3.3"
pytest-mock = "^3.14.0"
pytest-asyncio = "^0.24.0"
pre-commit = "^4.0.1"
types-beautifulsoup4 = "^4.12.0.20241020"
types-regex = "^2024.11.6.20241108"
types-chevron = "^0.14.2.20240310"
pytest-timeout = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
timeout = 60
asyncio_mode = "strict"
log_cli = true
log_cli_level = "INFO"
filterwarnings = [
    "ignore::DeprecationWarning:sklearn.utils.fixes:",
    "ignore::DeprecationWarning:pandas.core.common:",
    "ignore::pydantic.warnings.PydanticDeprecatedSince20:",
    "ignore::DeprecationWarning:importlib.resources._legacy:",
    "ignore::DeprecationWarning:litellm.utils:",
    "ignore:open_text is deprecated*:DeprecationWarning",
    "ignore:distutils Version classes are deprecated. Use packaging.version instead.*:DeprecationWarning",
    'ignore:configuration option "asyncio_default_fixture_loop_scope" is unset',
    "ignore:Valid config keys have changed in V2*:UserWarning"
]

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.isort]
profile = "black"
